refactor healthchecking for dev
